moved data loading to a custom react hook
