+++ src/App.tsx
@@ -1,39 +1,21 @@
-import React, { useEffect, useState } from 'react';
 import './App.css';
-import { fetchSheetData, IDatum } from './lib/data';
-import { useStorage } from './lib/storage';
+import { useLoadData } from './hooks/use-load-data';
 import {List} from "./List";
 import {ThemeToggle} from "./ThemeToggle";
 
 function App() {
-	const [ data, setData ] = useState<IDatum[]|null>(null);
-	const [ error, setError ] = useState<Error|null>(null);
-	const [ writeSavedList, readSavedList] = useStorage();
-
-	// On first render/mount, try to load data from localStorage first.
-	// If it doesn't exist then try fetching it.
-	useEffect(() => {
-		async function handleFetch() {
-			const [data, error] = await fetchSheetData();
-			setData(data);
-			setError(error);
-		}
-		try {
-			const savedList = readSavedList();
-			if (savedList && Array.isArray(savedList)) {
-				setData(savedList);
-			} else {
-				handleFetch();
-			}
-		} catch (error) {
-			setError(new Error("Problem reading localStorage."));
-		}
-	}, []);
+	const [data, error] = useLoadData();
 
 	// TODO: function that handles updating list items to be checked / unchecked.
-	// function updateListItem(id: number) {
+	// function updateListItem(index: number) {
 	// 	if (!data) return;
-	// 	setState()
+	// 	const datumCopy = { ...data[index] };
+	// 	datumCopy.checked = !datumCopy.checked;
+	// 	setData([
+	// 		...data.slice(0, index),
+	// 		datumCopy,
+	// 		...data.slice(index + 1)
+	// 	]);
 	// }
 
   return (
